add minor changes
